# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Product Service",
    "version": "1.0",
    "contact": {
      "name": "blueprint-roa-golang-grpc-gateway",
      "url": "http://localhost:3011",
      "email": "chat@robowealth.co.th"
    }
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/movie": {
      "get": {
        "operationId": "MovieService_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/api_v1MovieGetResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cinemaNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MovieService"
        ]
      },
      "delete": {
        "operationId": "MovieService_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/api_v1MovieDeleteResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MovieService"
        ]
      },
      "post": {
        "operationId": "MovieService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/api_v1MovieCreateResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api_v1MovieCreateRequest"
            }
          }
        ],
        "tags": [
          "MovieService"
        ]
      },
      "patch": {
        "operationId": "MovieService_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/api_v1MovieUpdateResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cinemaNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "dateUpdate",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeUpdate",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cinemaNoUpdate",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MovieService"
        ]
      }
    },
    "/api/v1/movie/list": {
      "get": {
        "operationId": "MovieService_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/api_v1MovieListResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cinemaNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "MovieService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "api_v1MovieCreateRequest": {
      "type": "object",
      "properties": {
        "movieName": {
          "type": "string",
          "title": "@inject_tag: validate:\"required\""
        },
        "date": {
          "type": "string",
          "title": "@inject_tag: validate:\"required\""
        },
        "time": {
          "type": "string",
          "title": "@inject_tag: validate:\"required\""
        },
        "cinemaNo": {
          "type": "string",
          "title": "@inject_tag: validate:\"required\""
        }
      }
    },
    "api_v1MovieCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "api_v1MovieDeleteResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "api_v1MovieGetResponse": {
      "type": "object",
      "properties": {
        "movieName": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "time": {
          "type": "string"
        },
        "cinemaNo": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "api_v1MovieListItem": {
      "type": "object",
      "properties": {
        "movieName": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "time": {
          "type": "string"
        },
        "cinemaNo": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "api_v1MovieListResponse": {
      "type": "object",
      "properties": {
        "item": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api_v1MovieListItem"
          }
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "total": {
          "type": "string",
          "format": "int64"
        },
        "totalPages": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "api_v1MovieUpdateResponse": {
      "type": "object",
      "properties": {
        "movieName": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "time": {
          "type": "string"
        },
        "cinemaNo": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}